+++ src/main/java/com/labygame/demo/HelloApplication.java
@@ -1,0 +1,27 @@
+package com.labygame.demo;
+
+import com.labygame.demo.personnage.CharacterState;
+import com.labygame.demo.personnage.Monster;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class HelloApplication extends Application {
+    @Override
+    public void start(Stage stage) throws IOException {
+        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello-view.fxml"));
+        Scene scene = new Scene(fxmlLoader.load(), 320, 240);
+        Monster myMonster = new Monster(50, CharacterState.NORMAL,"Demon",10,0,0);
+        stage.setTitle(myMonster.getName());
+        stage.setScene(scene);
+        stage.show();
+
+    }
+
+    public static void main(String[] args) {
+        launch();
+    }
+}
