+++ src/main/java/com/labygame/demo/items/Item.java
@@ -1,0 +1,5 @@
+package com.labygame.demo.items;
+
+import java.io.Serializable;
+
+public record Item(int bonus, ItemName name, String description) implements Serializable { }
