+++ src/main/java/com/labygame/demo/items/ItemName.java
@@ -1,0 +1,9 @@
+package com.labygame.demo.items;
+
+public enum ItemName {
+    ANTIDOTE, // to be use when the character is poison
+    HEALPOTION, // to be use when the character is sick
+    ENERGYDRINK, // to be use when the character is weary
+    KEY,
+    MAGICKEY;
+}
