+++ src/main/java/com/labygame/demo/personnage/CharacterState.java
@@ -1,0 +1,8 @@
+package com.labygame.demo.personnage;
+
+public enum CharacterState {
+    NORMAL,
+    SICK,
+    WEARY,
+    POISON
+}
