+++ src/main/java/com/labygame/demo/trayEnvironnement/Decor/Water.java
@@ -1,0 +1,21 @@
+package com.labygame.demo.trayEnvironnement.Decor;
+
+import com.labygame.demo.personnage.Hero;
+
+public class Water implements DecorInterface {
+
+    @Override
+    public boolean isDestroyed(){
+        return false;
+    }
+
+    @Override
+    public boolean canMove() {
+        return false;
+    }
+
+    @Override
+    public void hurtHero(Hero hero) {
+        hero.setHp(0);
+    }
+}
