+++ src/main/java/com/labygame/demo/trayEnvironnement/gameTray.java
@@ -1,0 +1,12 @@
+package com.labygame.demo.trayEnvironnement;
+
+public class gameTray {
+
+    private gameTrayPiece[][] gameBoard;
+
+    //Constructor
+    public gameTray(){
+        this.gameBoard = new gameTrayPiece[5][5];
+    }
+
+}
