+++ src/main/java/com/labygame/demo/trayEnvironnement/gameTrayPiece.java
@@ -1,0 +1,5 @@
+package com.labygame.demo.trayEnvironnement;
+
+public class gameTrayPiece {
+    //a remplir
+}
