--- src/main/java/labyGame/items/Item.java
@@ -1,5 +1,0 @@
-package labyGame.items;
-
-import java.io.Serializable;
-
-public record Item(int bonus, ItemName name, String description) implements Serializable { }
