--- src/main/java/labyGame/items/ItemName.java
@@ -1,9 +1,0 @@
-package labyGame.items;
-
-public enum ItemName {
-    ANTIDOTE, // to be use when the character is poison
-    HEALPOTION, // to be use when the character is sick
-    ENERGYDRINK, // to be use when the character is weary
-    KEY,
-    MAGICKEY;
-}
