--- src/main/java/labyGame/personnage/CharacterState.java
@@ -1,8 +1,0 @@
-package labyGame.personnage;
-
-public enum CharacterState {
-    NORMAL,
-    SICK,
-    WEARY,
-    POISON
-}
