--- src/main/java/labyGame/trayEnvironnement/Decor/Water.java
@@ -1,22 +1,0 @@
-package labyGame.trayEnvironnement.Decor;
-
-
-import labyGame.personnage.Hero;
-
-public class Water implements DecorInterface {
-
-    @Override
-    public boolean isDestroyed(){
-        return false;
-    }
-
-    @Override
-    public boolean canMove() {
-        return false;
-    }
-
-    @Override
-    public void hurtHero(Hero hero) {
-        hero.setHp(0);
-    }
-}
