--- src/main/java/labyGame/trayEnvironnement/gameTray.java
@@ -1,12 +1,0 @@
-package labyGame.trayEnvironnement;
-
-public class gameTray {
-
-    private gameTrayPiece[][] gameBoard;
-
-    //Constructor
-    public gameTray(){
-        this.gameBoard = new gameTrayPiece[5][5];
-    }
-
-}
