--- src/main/java/labyGame/trayEnvironnement/gameTrayPiece.java
@@ -1,5 +1,0 @@
-package labyGame.trayEnvironnement;
-
-public class gameTrayPiece {
-    //a remplir
-}
