--- src/main/java/lancer.java
@@ -1,5 +1,0 @@
-public class lancer {
-    public static void main (String[] args){
-        System.out.println("Main class");
-    }
-}
