+++ src/main/java/module-info.java
@@ -1,0 +1,11 @@
+module com.labygame.demo {
+    requires javafx.controls;
+    requires javafx.fxml;
+    requires static lombok;
+    requires org.jetbrains.annotations;
+    requires com.google.gson;
+    requires org.junit.jupiter.api;
+
+    opens com.labygame.demo to javafx.fxml;
+    exports com.labygame.demo;
+}
